+++ pkg/agent/collectors/events/service.go
@@ -35,12 +35,6 @@ type Service struct {
 }
 
 func (a *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluster, observabilityBundleVersion semver.Version) error {
-	// No-op if events collection is disabled at installation level
-	// TODO remove once the logging operator is gone
-	if !a.Config.Logging.EnableAlloyEventsReconciliation {
-		return nil
-	}
-
 	logger := log.FromContext(ctx)
 	logger.Info("alloy-events-service - ensuring alloy events is configured")
 
@@ -80,12 +74,6 @@ func (a *Service) ReconcileCreate(ctx context.Context, cluster *clusterv1.Cluste
 }
 
 func (a *Service) ReconcileDelete(ctx context.Context, cluster *clusterv1.Cluster) error {
-	// No-op if events collection is disabled at installation level
-	// TODO remove once the logging operator is gone
-	if !a.Config.Logging.EnableAlloyEventsReconciliation {
-		return nil
-	}
-
 	logger := log.FromContext(ctx)
 	logger.Info("alloy-events-service - ensuring alloy events is removed")